+++ src/main/java/com/example/portalegresso/model/entidades/Depoimento.java
@@ -27,7 +27,7 @@ public class Depoimento {
     @Id
     @Column(name="id_depoimento")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_depoimento;
+    private Long id_depoimento;
 
     @ManyToOne
     @JoinColumn(name="id_egresso")
