+++ src/main/java/com/example/portalegresso/model/repositorio/CursoRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.portalegresso.model.entidades.Cargo;
 
-public interface CursoRepositorio extends JpaRepository<Cargo,Integer>{
+public interface CursoRepositorio extends JpaRepository<Cargo,Long>{
 
 }
