+++ src/main/java/com/example/portalegresso/model/repositorio/EgressoRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.portalegresso.model.entidades.Egresso;
 
-public interface EgressoRepositorio extends JpaRepository<Egresso,Integer>{
+public interface EgressoRepositorio extends JpaRepository<Egresso,Long>{
     
 }
